cast unsigned int to int before negating it.
From gonsolo,
https://github.com/gonsolo/GraphicsGems/commit/5eff48b8d9f66c3ec0d23e4690a746ca63798679

+++ gemsiv/clahe.c
@@ -223,7 +223,7 @@ void MakeHistogram (kz_pixel_t* pImage, unsigned int uiXRes,
 		pImagePointer = &pImage[uiSizeX];
 		while (pImage < pImagePointer) pulHistogram[pLookupTable[*pImage++]]++;
 		pImagePointer += uiXRes;
-		pImage = &pImagePointer[-uiSizeX];	/* go to beginning of next row */
+		pImage = &pImagePointer[-(int)uiSizeX];	/* go to bdeginning of next row */
     }
 }
